requirements.txt
PyYAML~=6.0
dacite~=1.6.0
pip~=21.3.1
wheel~=0.37.1
Jinja2~=3.1.2
setuptools~=60.2.0
requests~=2.28.1
jsonschema~=2.6.0
chevron~=0.14.0